Add methods to calcualte slope and offset from two power points and two efficiencies
